+++ src/bootstrap/config.go
@@ -48,6 +48,7 @@ type Config struct {
 	FactorioIP              string `json:"factorio_ip,omitempty"`
 	FactorioAdminFile       string `json:"factorio_admin_file,omitempty"`
 	FactorioWhitelistFile   string `json:"factorio_whitelist_file,omitempty"`
+	FactorioBanFile         string `json:"factorio_ban_file,omitempty"`
 	ServerIP                string `json:"server_ip,omitempty"`
 	ServerPort              string `json:"server_port,omitempty"`
 	MaxUploadSize           int64  `json:"max_upload_size,omitempty"`
@@ -192,6 +193,12 @@ func (config *Config) loadServerConfig() {
 	if !filepath.IsAbs(config.FactorioWhitelistFile) {
 		config.FactorioWhitelistFile = filepath.Join(config.FactorioConfigDir, config.FactorioWhitelistFile)
 	}
+	if config.FactorioBanFile == "" {
+		config.FactorioBanFile = "server-banlist.json"
+	}
+	if !filepath.IsAbs(config.FactorioBanFile) {
+		config.FactorioBanFile = filepath.Join(config.FactorioConfigDir, config.FactorioBanFile)
+	}
 
 	if config.FactorioRconPort == 0 {
 		config.FactorioRconPort = randomPort()
@@ -225,6 +232,7 @@ func (config *Config) mapFlags(flags Flags) {
 	config.FactorioCredentialsFile = "./factorio.auth"
 	config.FactorioAdminFile = "server-adminlist.json"
 	config.FactorioWhitelistFile = "server-whitelist.json"
+	config.FactorioBanFile = "server-banlist.json"
 	config.ConsoleLogFile = filepath.Join(flags.FactorioDir, "factorio-server-console.log")
 	config.FactorioRconPort = flags.FactorioRconPort