+++ src/factorio/rcon.go
@@ -10,16 +10,15 @@ import (
 )
 
 func connectRC() error {
-	var err error
 	config := bootstrap.GetConfig()
 	rconAddr := config.ServerIP + ":" + strconv.Itoa(config.FactorioRconPort)
 	server := GetFactorioServer()
-	server.Rcon, err = rcon.Dial(rconAddr, config.FactorioRconPass)
+	console, err := rcon.Dial(rconAddr, config.FactorioRconPass)
 	if err != nil {
 		log.Printf("Cannot create rcon session: %s", err)
 		return err
 	}
-	server.SetRconConnected(true)
+	server.SetRCON(console)
 	log.Printf("rcon session established on %s", rconAddr)
 
 	return nil